+++ src/main/java/com/quantest/quantestbe/domain/stock/dto/StockRankingDto.java
@@ -1,5 +1,7 @@
 package com.quantest.quantestbe.domain.stock.dto;
 
+import java.math.BigDecimal;
+
 import com.querydsl.core.annotations.QueryProjection;
 
 import lombok.Getter;
@@ -8,12 +10,12 @@ import lombok.Getter;
 public class StockRankingDto {
 	private Long stockId;
 	private String stockName;
-	private double chartChangePercentage;
+	private BigDecimal chartChangePercentage;
 	private int chartClose;
 	private char recordDirection;
 
 	@QueryProjection
-	public StockRankingDto(Long stockId, String stockName, double chartChangePercentage, int chartClose, char recordDirection) {
+	public StockRankingDto(Long stockId, String stockName, BigDecimal chartChangePercentage, int chartClose, char recordDirection) {
 		this.stockId = stockId;
 		this.stockName = stockName;
 		this.chartChangePercentage = chartChangePercentage;
